AB#329858- fs fixes

+++ app/src/main/java/com/delhivery/axle/ui/fastag/FastagRechargeViewModel.kt
@@ -33,7 +33,7 @@ class FastagRechargeViewModel @Inject constructor(
     }
 
 
-    fun rechargeFastag(fastagId: String, amount: Int, fastagStatus: String?, fastagBalance: String?) {
+    fun rechargeFastag(fastagId: String, amount: Double, fastagStatus: String?, fastagBalance: String?) {
         showProgress()
         val request = FastagRechargeRequest(fastagId = fastagId, rechargeAmount = amount, fastagStatus = fastagStatus, fastagBalance = fastagBalance)
